+++ CHANGELOG.rst
@@ -14,7 +14,7 @@ v0.4.0 (02/02/2023)
 **Changed**
 
 - Collection can now return up to 3 tomograms, returns paged object for :code:`/tomograms` (renamed from :code:`/tomogram`)
-
+- Proposal search also searches through title
 
 ++++++++++
 v0.3.2 (02/02/2023)